third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/computed/computed-expected.txt

This is a testharness.js-based test.
FAIL Computed StylePropertyMap contains every CSS property assert_equals: expected 296 but got 294
PASS Computed StylePropertyMap contains CSS property declarations in style rules
PASS Computed StylePropertyMap contains custom property declarations in style rules
PASS Computed StylePropertyMap contains CSS property declarations in inline styles
PASS Computed StylePropertyMap contains custom property declarations in inline rules
PASS Computed StylePropertyMap contains computed values and not resolved values
PASS Computed StylePropertyMap is live
Harness: the test ran to completion.
